requirements.txt
ollama>=0.1.0
sounddevice>=0.4.6
numpy>=1.21.0
scipy>=1.9.0
noisereduce>=3.0.0
openai-whisper>=20231117
piper-tts>=1.2.0
onnxruntime>=1.16.0
torch>=2.0.0
transformers>=4.30.0
soundfile>=0.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
#all the requirements for the project, must download all of them before running
